% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbeUtils-package.R
\docType{package}
\name{cbeUtils-package}
\alias{cbeUtils}
\alias{cbeUtils-package}
\title{cbeUtils: Various utility functions for WDFW and FRAM team work}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cbedwards-dfw/cbeUtils}
  \item \url{https://cbedwards-dfw.github.io/cbeUtils/}
  \item Report bugs at \url{https://github.com/cbedwards-dfw/cbeUtils/issues}
}

}
\author{
\strong{Maintainer}: Collin Edwards \email{Collin.Edwards@dfw.wa.gov} (\href{https://orcid.org/0000-0002-4937-5159}{ORCID})

}
\keyword{internal}
