% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_time_blocks.R
\name{consolidate_time_blocks}
\alias{consolidate_time_blocks}
\title{Look for breaks in sequences of time blocks}
\usage{
consolidate_time_blocks(.data = NULL, start_col = "start", end_col = "end")
}
\arguments{
\item{.data}{Dataframe. Optional; if not provided, instead use system clipboard (assumes it contains a copy of a 2-column excel block)}

\item{start_col}{Name of column containing start dates. Defaults to "start". Unused if \code{.data} is not provided.}

\item{end_col}{Name of column containing end dates. Defaults to "start". Unused if \code{.data} is not provided.}
}
\value{
Tibble with start and end date of each non-contiguous period, and length in days
}
\description{
Developed to help work with fishing regulation data. \code{consolidate_time_blocks} takes any number of rows, with a column of start and end dates, and consolidates all contiguous periods of dates into a single one (and so each row of the output represents a period with one or more days break with the last one). Moreover, this function uses the {clipr} package to use the system clipboard. So you can copy a block of dates from excel (like in the screenshot) and then run the function with no arguments to have it pull in the copied cells and run the calculations.
}
\examples{

library(tidyr)
df = tribble(
~ start, ~end_date,
"2024-8-1", "2024-8-5",
"2024-8-6", "2024-8-10",
"2024-8-16", "2024-8-30"
)
df |> consolidate_time_blocks(end_col = "end_date") #column name for start column is the default, "start"
}
